+++ packages/react-core/src/evm/hooks/chain-hooks.ts
@@ -1,13 +1,13 @@
 import { useMemo } from "react";
-import { ThirdwebSDKProviderProps } from "../../evm";
-import { updateChainRPCs } from "../utils/updateChainRpcs";
+import { updateChainRPCs } from "../../core/utils/updateChainRpcs";
+import { ThirdwebSDKProviderProps } from "../providers/types";
 
 export function useUpdateChainsWithApiKeys(
   supportedChains: NonNullable<ThirdwebSDKProviderProps["supportedChains"]>,
   activeChain: ThirdwebSDKProviderProps["activeChain"],
-  thirdwebApiKey: string,
-  alchemyApiKey: string,
-  infuraApiKey: string,
+  thirdwebApiKey?: string,
+  alchemyApiKey?: string,
+  infuraApiKey?: string,
 ) {
   const keys = useMemo(
     () => ({
